Conflicts:
	apps/opencs/model/doc/document.cpp
	apps/opencs/model/doc/documentmanager.cpp
	components/CMakeLists.txt
	components/bsa/resources.cpp
	components/nif/data.hpp
	components/nif/node.cpp
	components/nifogre/mesh.hpp
	components/nifogre/ogrenifloader.cpp
	components/nifogre/particles.cpp

+++ apps/esmtool/esmtool.cpp
@@ -461,7 +461,7 @@ int clone(Arguments& info)
     for (Stats::iterator it = stats.begin(); it != stats.end(); ++it)
     {
         name.val = it->first;
-        float amount = it->second;
+        int amount = it->second;
         std::cout << std::setw(digitCount) << amount << " " << name.toString() << "  ";
 
         if (++i % 3 == 0)
